Fix shipping records w/o shopId (#2791)
* When creating shipping records, create one for each shop

* Return an empty object if cart has no items

* Fix grammatical errors in test description

* Update copyToCart to work with by shop shipping records

* Add billing/shipping shopIds to fix stripe test

* Break copyCartToOrder to it's own file

* Add shopId to schema

* Set selected method on all shipping records

* Only show shipping methods associate with the primary shop

* Check rates against schema

* Remove blackbox from Cart.shipping so it will actually validate

* Validate shipping methods

* Put blackbox back

* simple-schema not working, remove

* only show shipping methods from primaryShopId

* Lint fixes

* Assign items to shipping records based on shopId

* Assign shopIds when adding shipping address

* Only show the address when it has data

* Lint fix

* FIx tests
