+++ imports/plugins/core/shipping/client/templates/checkout/shipping.js
@@ -9,13 +9,15 @@ import { Cart } from "/lib/collections";
 function cartShippingQuotes(currentCart) {
   const cart = currentCart || Cart.findOne();
   const shipmentQuotes = [];
-
+  const primaryShopId = Reaction.getPrimaryShopId();
   if (cart) {
     if (cart.shipping) {
       for (const shipping of cart.shipping) {
         if (shipping.shipmentQuotes) {
           for (const quote of shipping.shipmentQuotes) {
-            shipmentQuotes.push(quote);
+            if (shipping.shopId === primaryShopId) {
+              shipmentQuotes.push(quote);
+            }
           }
         }
       }
@@ -29,11 +31,11 @@ function cartShippingQuotes(currentCart) {
 function cartShipmentMethods(currentCart) {
   const cart = currentCart || Cart.findOne();
   const shipmentMethods = [];
-
+  const primaryShopId = Reaction.getPrimaryShopId();
   if (cart) {
     if (cart.shipping) {
       for (const shipping of cart.shipping) {
-        if (shipping.shipmentMethod) {
+        if (shipping.shipmentMethod && shipping.shopId === primaryShopId) {
           shipmentMethods.push(shipping.shipmentMethod);
         }
       }
