+++ lib/collections/schemas/shipping.js
@@ -236,6 +236,9 @@ export const Shipment = new SimpleSchema({
     label: "Shipment Id",
     autoValue: schemaIdAutoValue
   },
+  shopId: {
+    type: String
+  },
   paymentId: {
     type: String,
     label: "Payment Id",
