+++ lib/collections/transform/cartOrder.js
@@ -222,14 +222,17 @@ export const cartOrderTransform = {
    * @return {Object} Dict of shopIds with an array of items from that shop that are present in the cart/order
    */
   getItemsByShop() {
-    return this.items.reduce((itemsByShop, item) => {
-      if (!itemsByShop[item.shopId]) {
-        itemsByShop[item.shopId] = [item];
-      } else {
-        itemsByShop[item.shopId].push(item);
-      }
-      return itemsByShop;
-    }, {});
+    if (this.items) {
+      return this.items.reduce((itemsByShop, item) => {
+        if (!itemsByShop[item.shopId]) {
+          itemsByShop[item.shopId] = [item];
+        } else {
+          itemsByShop[item.shopId].push(item);
+        }
+        return itemsByShop;
+      }, {});
+    }
+    return {};
   },
   /**
    * Returns an array of payment methods, normalized
